Set up the database as part of the provisioning process.
I find it kinda fiddly to copy `database.sample.yml` and edit things
according to my local install. If we are going to use Vagrant, then
might as well get the provisioning script to do that as well.

A nice side-effect is that, if the Vagrant box was brought up OK,
all the tests pass immediately.